backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.3.0
bcrypt==4.1.3
python-dotenv==1.0.1
gunicorn==22.0.0

# Dev / CI dependencies
pytest==8.2.0
pytest-cov==5.0.0
flake8==7.0.0
